+++ src/app/post/[post_id]/components/DeleteConfirmModal/DeleteConfirmModal.styles.ts
